+++ extensions-leaflet-map.php
@@ -5,7 +5,7 @@
  * GitHub Plugin URI: https://github.com/hupe13/extensions-leaflet-map-github
  * Primary Branch:    main
  * Description:       Extensions for the WordPress plugin Leaflet Map Github Version
- * Version:           3.4.1-230313
+ * Version:           3.4.1-230315
  * Requires PHP:      7.4
  * Author:      hupe13
  * License:     GPL v2 or later
@@ -62,6 +62,7 @@ include_once LEAFEXT_PLUGIN_DIR . '/php/extramarker.php';
 include_once LEAFEXT_PLUGIN_DIR . '/php/choropleth.php';
 
 include_once LEAFEXT_PLUGIN_DIR . '/php/zoomhome.php';
+//include_once LEAFEXT_PLUGIN_DIR . '/php/zoomhome-a.php';
 include_once LEAFEXT_PLUGIN_DIR . '/php/layerswitch.php';
 include_once LEAFEXT_PLUGIN_DIR . '/php/providers.php';