+++ changes.md
@@ -1,4 +1,4 @@
-* revision of hover: bugs fixed, many new options (Bug: does not work together with zoomhomemap!!!)
+* revision of hover: bugs fixed, many new options
 * leaflet-optiongroup: leaflet-polyline, -circle, -line
 * leaflet-search: leaflet-polyline, -circle, -line
 * leaflet-search: css for iPhone